build: change tsc target to es2017 instead of es2020
There's an issue in older webviews trying to run the generated es2020 javascript. Specifically the 'Optional chaining (?.)' syntax. After rebuilding using es2017 the older webview doesn't have any issue running the script anymore.

I don't know if this was the best solution, or if there's a way to transpile/polyfill in the consuming project instead. I tried a couple of different configuration changes using babel in kidsloop-live-frontend but wasn't able to make it work.