fix: SFU-61 (#14)
* fix(sfu): add delay before replacing track to prevent ping-ponging the device track when pausing

When you pause and unpause quickly, you can cause the track to be continuously replaced. This adds a
timer to only replace the track after it has been paused for 5 seconds.

* refactor(sfu, track, tracksender): extract track logic from SFU file

* fix(sfu, track, webrtc): pause consumer on producer pause, delay to pause state change on consumer

Failing to pause the consumer when the producer rapidly switches pause states causes issues on the
consuming side.  I've added a delay if the pause state is switching rapidly to actually pause the
consumer, this seems to resolve SFU-61

fix #SFU-61