refactor: Bug/transports (#19)
* fix(network): use mutex when accessing transports, create new transport on disconnect

Also shut down producers and consumers when the underlying transport die

* refactor: add mutex to trackSender state

* chore(track, tracksender): rebase onto main

* style(network, rtccontext, memoize): refactor to improve readability of complex code paths

* fix(track): fix pause delay when coming out of long pause state

You can no longer spam the unpause/pause after being paused for a while

* fix(sfu): don't clear retries unless we have successfully connected to an SFU

don't clear retries on reaching max attempts

* refactor(sfutypes): move types from sfu.ts into their own file

removes some of the bloat in the sfu.ts file and makes it easier to navigate

* style(sfu.ts): make retry parameters modifiable in constructor, more consistent function notation

* refactor(sfu, track, webrtc): despaghettify track

The SFU should respond to changes in the track state instead of passing down its member functions
into other objects and allowing tracks to call them directly.  This decouples the track and SFU.

* refactor(sfu): guard against multiple simultaneous pause requests to maintain old behavior

previously the track `pause` method waited on the result of the request.  Now that the track is
decoupled, I have the SFU wait to send requests of the same type until the previous one has resolved

* refactor(network, react): move to more event driven architecture

By removing shared methods, objects are no longer extremely tightly coupled to each other.  This
allows me to more easily replace the objects in the event of a failure, and clarifies many sections
of logic since the methods can be used clos

* build(package): remove prettier

* feat(network, react): further separate transports from their parent objects, adjust hooks

WIP

* refactor(network, react): get consumers/producers working with new event driven architecture

removed a deadlock when creating a consumer due to holding the pause lock then trying to pause

* fix(webrtc): fix default state when there is no track on the tracksender

the default state was set to unpaused.  However, this forced you to only be able to try to pause
your track, and never let you unpause it since the initial button believed you were unpaused.

* refactor(sfu): create createTrackLock decorator

* refactor(sfu): create producerTransportLock decorator

* refactor(sfu): move producer transport acquisition up one function, stack decorators

* refactor(sfu): add consumerTransportLock decorator

* refactor(sfu): apply createTrackLock to consumeTrack

* refactor(sfu): add decorator for acquiring the pauseLock

* refactor(sfu): apply pauseLock decorator to pauseGlobally

* refactor(sfu): relax createTrackLock when just wanting to read a value

wait for the lock to unlock, but don't then acquire it

* refactor(sfu): fix printing of lock releases, use decorator in createProducer

* refactor(trackSender): Add decorator for statechangelock access

* refactor(network): toggle lock log printing

* refactor(sfu, sfuTypes): extract onProduce method

* refactor(network, react): don't request screen multiple times

* fix(react): fix screenshare

name needed an extra "-video" for the screenshare sender type

* refactor(network, react): remove verbose logging

* style(networktransport): add decorator for networkTransport wsLock